Create point.h
Point.h is the header file for the point class. It contains the definition of the point class and the documentation of the methods included in the point class.

This is a part of a CS 202 project where I had to create a point class and perform certain tasks (shifts, rotations) through Object Oriented Programming.