Create bag.h
bag.h is the header file for a bag class (adopted from MAIN/SAVITCH C++ DATA STRUCTURES + OBJECTS) that works as a partially filled array to represent a bag.

As is the standard, this header file will actually contain documentation of the bag class methods as well as the actual definition of the bag class within the namespace Jarrett_Greig. A macro guard is also included in order to avoid defining the class more than once.