Create PythonDungeonGame.py
This is a game I developed quickly in Python within 24 hours of picking up the language earlier this month. It functions as your typical text-based command-in-results-out game. From design to implementation, this was a good exercise in decision making and I feel like the idea of using functions as "rooms" is an easy way to tell the game which rooms are adjacent to each other.